+++ quirrel/package.json
@@ -53,7 +53,9 @@
     "basic-auth": "^2.0.1",
     "chalk": "^4.1.0",
     "commander": "^6.1.0",
+    "conditional-type-checks": "^1.0.5",
     "cron-parser": "^2.16.3",
+    "cross-fetch": "^3.0.6",
     "fastify": "^3.9.2",
     "fastify-basic-auth": "^1.0.1",
     "fastify-blipp": "^3.0.0",
@@ -63,11 +65,14 @@
     "fastify-websocket": "^2.0.7",
     "ioredis": "^4.19.2",
     "ioredis-mock": "^5.1.0",
+    "ms": "^2.1.3",
     "open": "^7.3.0",
     "parse-duration": "^0.4.4",
     "pino": "^6.7.0",
+    "secure-e2ee": "^0.2.0",
     "secure-webhooks": "^0.2.0",
     "uuid": "^8.3.0",
-    "ws": "^7.3.1"
+    "ws": "^7.3.1",
+    "zod": "^1.11.11"
   }
 }
