--- quirrel/src/client/jest.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  testPathIgnorePatterns: [
-    "/dist"
-  ]
-};
