--- quirrel/src/client/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@quirrel/client",
-  "version": "0.10.0",
-  "description": "Quirrel Client",
-  "main": "dist/src/index.js",
-  "files": [
-    "dist/src/"
-  ],
-  "scripts": {
-    "build": "tsc",
-    "prepack": "npm run build",
-    "test": "jest --runInBand --forceExit"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/quirrel-dev/quirrel.git"
-  },
-  "author": {
-    "email": "info@simonknott.de",
-    "name": "Simon Knott",
-    "url": "https://simonknott.de"
-  },
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/quirrel-dev/quirrel/issues"
-  },
-  "homepage": "https://github.com/quirrel-dev/quirrel#readme",
-  "devDependencies": {
-    "@types/ioredis": "4.17.7",
-    "@types/ms": "^0.7.31",
-    "@types/node": "^14.11.2",
-    "delay": "^4.4.0",
-    "ioredis": "^4.19.2",
-    "jest": "^26.5.3",
-    "prettier": "^2.1.2",
-    "quirrel": "file:../api",
-    "ts-jest": "^26.4.1",
-    "typescript": "^4.0.3"
-  },
-  "dependencies": {
-    "conditional-type-checks": "^1.0.5",
-    "cross-fetch": "^3.0.6",
-    "ms": "^2.1.2",
-    "secure-e2ee": "^0.1.1",
-    "secure-webhooks": "^0.2.0",
-    "zod": "^2.0.0-beta.29"
-  }
-}
