+++ quirrel/src/client/test/ms.test.ts
@@ -1,5 +1,5 @@
 import { QuirrelClient } from "../src";
-import { runQuirrel } from "quirrel";
+import { runQuirrel } from "../../api";
 import type { AddressInfo } from "net";
 import * as http from "http";
 import delay from "delay";
